Added Exit/Continue buttons
Implementation for game scenes Magnetometer, Spectrometer, and Grav. Science, for the demo; Magnetometer has a button that takes you to the Orbit minigame, Magnetometer and Grav Sci have buttons that take you to the next minigame (assuming Magnetometer -> Imager -> Grav. Sci. -> Spectrometer), and all 3 have buttons that take you to the title screen